Moving session config from internal to public
