+++ driver/src/main/java/org/neo4j/driver/internal/SessionFactory.java
@@ -20,6 +20,7 @@ package org.neo4j.driver.internal;
 
 import java.util.concurrent.CompletionStage;
 
+import org.neo4j.driver.SessionConfig;
 import org.neo4j.driver.internal.async.NetworkSession;
 
 public interface SessionFactory
