+++ driver/src/test/java/org/neo4j/driver/internal/DirectDriverBoltKitTest.java
@@ -56,8 +56,8 @@ import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.neo4j.driver.internal.SessionConfig.builder;
-import static org.neo4j.driver.internal.SessionConfig.forDatabase;
+import static org.neo4j.driver.SessionConfig.builder;
+import static org.neo4j.driver.SessionConfig.forDatabase;
 import static org.neo4j.driver.internal.logging.DevNullLogging.DEV_NULL_LOGGING;
 import static org.neo4j.driver.Values.parameters;
 import static org.neo4j.driver.util.StubServer.INSECURE_CONFIG;
