+++ driver/src/test/java/org/neo4j/driver/stress/AbstractRxQuery.java
@@ -22,7 +22,7 @@ import org.neo4j.driver.AccessMode;
 import org.neo4j.driver.Driver;
 import org.neo4j.driver.reactive.RxSession;
 
-import static org.neo4j.driver.internal.SessionConfig.builder;
+import static org.neo4j.driver.SessionConfig.builder;
 
 public abstract class AbstractRxQuery<C extends AbstractContext> implements RxCommand<C>
 {
